--- Homework-Async/EjbTest/nbproject/private/private.properties
@@ -1,7 +1,0 @@
-deploy.ant.properties.file=C:\\Users\\J\u00f3zsi\\AppData\\Roaming\\NetBeans\\8.1\\config\\GlassFishEE6\\Properties\\gfv31004099329.properties
-j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/GlassFish_Server/glassfish/domains/domain1
-j2ee.server.home=C:/GlassFish_Server/glassfish
-j2ee.server.instance=[C:\\GlassFish_Server\\glassfish;C:\\GlassFish_Server\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=C:/GlassFish_Server
-user.properties.file=C:\\Users\\J\u00f3zsi\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
